Add per-mount uid/gid override
The new inheritid filesystem property can be set to "uid", "gid" or "all".
It is interpreted by the zfs.mount helper to set the temporary mount options
indicating the uid and/or gid of all files and directories in the filesystem
should be overridden to match that of, respectively, the uid or gid of the
directory on which the filesystem is being mounted.

This is useful for pools on removable media which are transported between
systems without synchronized user and group ids.

Additionally, the zfs.mount helper will interpret any uid=<id> or gid=<id>
mount options it is passed on the command line the ownership or group
override to be manually specified to something other than those of the
directory on which the filesystem is being mounted.

A uid override causes all existing files to appear to be owned by the
specified user id.  It also causes newly-created files to inherit the
uid of their containing directory.  Group ids are handled similarly.

Since posix ACLs are specified as integer uid/gid values, they are
incompatible with this feature.  Using either uid or gid override has
the effect of setting acltype=noacl.