Fix zdb(8) hang when running zloop
When attempting to reconstruct a split block from an indirect vdev
in the IO pipeline use the number of segments to assess if it is
reasonable to attempt all combinations.  If it is do so.  Otherwise
randomly sample the possible combination space under the assumption
that most of the segments are valid.

An alternate strategy which has been left as a future optimization
is to make an initial pass over the segment copies to identify
duplicate children.  This potentially could drastically reduce
the search space although these still no assurance it would be
sufficient.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>