Change parent/child locking order
This was taken as a subset of a8b2e30 (Support re-prioritizing
asynchronous prefetches) and prevents the deadman zio iterator from
deadlocking with child removal.

NOTE: After a brief back-and-forth with Tom Caputi and subsequent
follow-up by Brian Behlendorf in PR #6926, I added the successor commit
to this one.  They should be considered to be part of the same fix.