Only call spa_man_trim_taskq_destroy() when needed
Long zloop runs were occasionally hitting the
"ASSERT(spa->spa_man_trim_taskq != NULL)" in spa_man_trim_taskq_destroy().
It's not clear to me how this was happening because the only place
"spa->spa_man_trim_taskq" is cleared is in spa_man_trim_taskq_destroy()
itself which is only called from the (single) spa_async_thread() or
from spa_unload().

To that end, this commit adds a non-NULL check in spa_async_thread()
analagous to the tests which were added when support for stopping/starting
the auto trim taskqs were added to spa_async_thread().

NOTE: Yes, that means I consider this to be a band-aid.